+++ de.metas.adempiere.adempiere/base/src/main/java/de/metas/product/ResourceId.java
