+++ de.metas.material/planning/src/main/java/de/metas/material/planning/pporder/PPOrderPojoConverter.java
@@ -22,6 +22,7 @@ import de.metas.material.event.pporder.MaterialDispoGroupId;
 import de.metas.material.event.pporder.PPOrder;
 import de.metas.material.event.pporder.PPOrderLine;
 import de.metas.organization.OrgId;
+import de.metas.product.ResourceId;
 import de.metas.util.Services;
 import lombok.NonNull;
 
@@ -67,7 +68,7 @@ public class PPOrderPojoConverter
 				.dateStartSchedule(asInstant(ppOrderRecord.getDateStartSchedule()))
 				.docStatus(DocStatus.ofCode(ppOrderRecord.getDocStatus()))
 				.orgId(OrgId.ofRepoId(ppOrderRecord.getAD_Org_ID()))
-				.plantId(ppOrderRecord.getS_Resource_ID())
+				.plantId(ResourceId.ofRepoId(ppOrderRecord.getS_Resource_ID()))
 				.ppOrderId(ppOrderRecord.getPP_Order_ID())
 				.productDescriptor(productDescriptorFactory.createProductDescriptor(ppOrderRecord))
 				.productPlanningId(ppOrderRecord.getPP_Product_Planning_ID())
